#2 Abstracted hiddenNavigationBarStyle
