+++ TheadDarts/View/Blocks/PlayerUnitView.swift
@@ -9,26 +9,27 @@
 import SwiftUI
 
 struct PlayerUnitView<Score: DartScore>: View {
-    @ObservedObject var playerUnitVM: PlayerUnitViewModel<Score>
+  @ObservedObject var playerUnitVM: PlayerUnitViewModel<Score>
     
-    let startPlayerNameEditing: (String, @escaping (String) -> ()) -> ()
+  let startPlayerNameEditing: (String, @escaping (String) -> ()) -> ()
         
-    var body: some View {
-        Button(action: { withAnimation {
-            self.startPlayerNameEditing(self.playerUnitVM.name, self.playerUnitVM.update(nameTo:)) }
-        }) {
-            VStack(spacing: 0) {
-                Text(self.playerUnitVM.name)
-                    .font(.headline)
-                if playerUnitVM.showPoints {
-                    Text(String(self.playerUnitVM.points))
-                        .font(.title)
-                }
-            }
-            .lineLimit(1)
-            .frame(maxWidth: .infinity, alignment: .center)
+  var body: some View {
+    Button(action: { withAnimation {
+      self.startPlayerNameEditing(self.playerUnitVM.name, self.playerUnitVM.update(nameTo:)) }
+    }) {
+      VStack(spacing: 0) {
+        Text(self.playerUnitVM.name)
+          .font(.headline)
+        if playerUnitVM.showPoints {
+          Text(String(self.playerUnitVM.points))
+            .font(.title)
         }
+      }
+      .lineLimit(1)
+      .frame(maxWidth: .infinity, alignment: .center)
     }
+    .foregroundColor(Color.select(.primary))
+  }
 }
 
 struct PlayerUnitView_Previews: PreviewProvider {
